Add CONTRIBUTING guidelines for all tektoncd projects ✔️
This takes the contributing guidelines from
https://github.com/tektoncd/pipeline/blob/master/CONTRIBUTING.md
and generalizes (and updates) them to apply to all projects in the
tektoncd org.

Following up to this, I will remove most of the content of
https://github.com/tektoncd/pipeline/blob/master/CONTRIBUTING.md
and link to this community repo instead.

Some new things:

* The guidelines on becoming an OWNERS are adapted from https://github.com/knative/docs/blob/master/contributing/ROLES.md#approver
* I changed the OWNERS file to contain the members of the governing
  board
* I arbitrarily said that 50% of the governing board should review
  PRs in this repo that change policy

Fixes #3 Fixes #1

p.s. Shout out to the person who came up to me yesterday and was so sad
that we had no contributing guidelines b/c he went straight to the
community repo and there was nothing there!! Very motivating moment for
me XD