Move working group to Wednesdays 📅
In a recent working group meeting, it was brought up that the CDF TOC
meeting overlapped with the Tekton Working group meeting (which has been
on Tuesdays at 9am PST) so I sent around a form and the most popular
option to move the meeting to is Wednesdays at 9am PST so let's try that
for a while.

This commit updates the info and posts a new link to the calendar event
which should now be on Wednesdays.

Relevant thread on the mailing list:
https://groups.google.com/forum/#!topic/tekton-dev/MlUfUJU-7N8