Add link to shared doc drive 📁
This shared drive has been where we've been putting meeting group
recordings all this time but we haven't yet been using it for other
docs. Folks have been asking more and more recently for something like
this so now the drive should be editable by folks on the mailing list
and we can try to making design docs and meeting notes available here
when possible.