Add calendar and triggers sync to docs 📅
We've been having some ad-hoc meetings about triggers that have been
becoming more regular as we've found more in the design to discuss, and
more folks are interested in joining these meetings, so they are now
available on the public Tekton calendar!

Also added a link to the calendar itself and to a `#meetings` channel in
slack where people can announce more ad-hoc meetings that they'd like to
invite community folks to.