Triggers meeting is now every week! ⚡
As discussed in the working group today, and proposed by @dibyom,
we want to have more frequent but less long meetings. Especially if we
want to have more frequent releases. This is also more like what the
other working groups (that have started since) are doing.

I don't even remember the meeting being on monday! hah!