Move working group to 1:00pm EST on thursdays 📅
Survey for the working group time was tied between this and the same
time on Tuesdays. I have the CDF outreach meeting at 1pm tuesday and
Dibyo pointed out the triggers meeting is already on Tuesdays (and some
other working groups too) so staking out Thursday for this one.

And fix the calendar link, thanks #67