Add requirements for joining the Tekton org 👪
Previously we had no requirements beyond asking to join, but recently
the new governing board (https://github.com/tektoncd/community/pull/90)
discussed and tried to identify:

* What happens if you join (you can lgtm + you get ok-to-test)
* Why you would want this (probably only if you are regularly
  contributing)

The lgtm is something we want to be careful
around since this could result in arbititrary code in our released
binaries.

Even with ok-to-test we want to exercise caution because this means
running arbitrary code on our infrastructure with test priviledges.

The requirements in this commit are meant to attempt to demonstrate that
this person has been shown to be a contributor who needs and can
responsibly use these abilities.