Add info about EasyCLA 👩‍⚖️
We are switching Tekton over from using the Google CLA to using the CDF
CLA. This commit adds some docs about how it works. It is a bit
aspirational since at the moment I think I might be the only governing
board member who has access to admin the CLA however I have put in a
request to get everyone access.

I also thought CONTRIBUTING.md might be a better place for this info
however the CONTRIBUTING.md in this repo is only for this repo so I then
thought process.md might be a better place for org wide info. I also
wasn't sure if plumbing was a better place but this seems in between
since external contributors might want this info also.