Add contributor and reviewer expectations
These started out as just being for Tekton Pipelines, but @dibyom
pointed out they make sense for triggers too, so let's see if we want to
adopt these for all the tektoncd projects or if we just want to
constrain it to triggers + pipelines.

The goal of these guidelines is to make sure that contributors
understand what is expected in their submissions and also so that a
reviewer (especially if someone wants to join and start reviewing who
hasn't reviewed as part of this project before) will know when they
approve + lgtm what they are expected to have checked.