TEP-0007: Conditions Beta
Conditions is a CRD used to specify a criteria to determine whether or
not a Task executes. When other Tekton resources were migrated to beta,
it remained in alpha because it was missing features needed by users.
After analyzing the feature requests and discussing with users, we have
identified that the most critical gaps in Conditions are simplicity,
efficiency and skipping. We want to address these gaps so that it can
work well with the other Pipeline resources and users can count on its
stability.