Add TEP for new (sub-)metrics
This is a barebone TEP for adding new metrics as
suggested by the TEP template. At this point it
has only the summary, motivation, and a testing plan
to get aggreement on the general path.

I will iterate more to add the proposed submetrics
and what they measure and where should they be
measured in the code in subsequent PRs.