Propose optional workspaces feature as TEP
Workspaces declared in Tasks are currently always required: a TaskRun
cannot elect to omit a Workspace even if the data produced or consumed
on that Workspace is irrelevant to the user's needs.

This TEP proposes a new boolean `optional` field for Workspace
declarations in Tasks. When the optional field is true a TaskRun
may choose to omit a Workspace binding, resulting in no filesystem
being mounted at that location. This feature would support a couple of
use cases:

- credentials could be optionally provided. e.g. git-clone could accept
a .ssh directory for cloning private repos
- catalog tasks could be written to accept either an optional
pipelineresource OR an optional workspace at the same location on
disk, resulting in a Task that is more flexible to user need
- catalog tasks could be written to support both creds-init and explicit
credentials-via-workspace.
- tasks could be written to optionally take some extra configuration,
like a lint config that changes some default errors into warnings
- tasks could be written to generate some optional data, such as a manifest
of files copied to external storage.
- tasks would be able to conditionally skip some amount of work if an
optional workspace is not provided by a TaskRun: there'd be no need
to generate verbose test output to a file if no workspace was passed in
to receive that verbose test result file.