Add a TEP for Step and Sidecar Workspaces
Prior to this commit Workspace types have been limited to Tasks, TaskRuns, Pipelines and PipelineRuns.
There are known limitations with this:

- Problems with our credentials system are similarly present in our Workspace system: sensitive
data is exposed to every single Step in a Task instead of only those Steps that actually make
use of the data.

- Sidecars do not have a "blessed" path to accessing a Workspace. Instead they must rely on
a hack by adding the $(workspaces.<name>.volume) variable to the `name` field of a `volumeMount`.

The TEP in this commit is a proposal that describes the problem and covers a number of user stories
that should be tackled.