TEP 0007: Conditions Beta Skipping
Previously, we'd proposed to adding `continueAfterSkip` field which
defaults to false and can be set to true to execute ordering-dependent
tasks

We want to update that proposal because:
- using string aliases instead of booleans is preferable because,
as described in the Kubernetes API policy, "Think twice about bool
fields. Many ideas start as boolean but eventually trend towards a
small set of mutually exclusive options. Plan for future expansions by
describing the policy options explicitly as a string type alias"
- that naming was misleading

So, we propose using a `whenSkipped` field which defaults to `skipBranch`
and users can set to `runBranch` to execute its branch that's made up of
ordering-dependent tasks only

Further discussed in https://github.com/tektoncd/pipeline/pull/3176