Create teams of collaborators
Create a collaborators team for every repo. Individuals in the
collaborators team have read access granted, same as the current
level of access that is provided org wide to all individuals that
belong to the org.

Every individual with read (or more) access to a repo is considered
a collaborator by GitHub; the list of collaborator is what the
/lgtm prow plugin uses to determin if an individual is allowed
to /lgtm a patch.

Ultimately, these team of collaborators aim to give more
granularity in how access to /lgtm is granted, from the whole
GitHub org to project specific, which is more in line with the
spirit of the lgtm plugin:
https://github.com/kubernetes/test-infra/blob/master/prow/plugins/approve/approvers/README.md#lgtm-label

The teams of collaborators have been created using a script
that looks for individuals that contributed 5+ commits ever to
a specific repo and who belong to the Tekton org. Hopefully this
approach will ensure that noone is affected in their daily work
by this change.

Until the organization default access setting is flipped from
read to none, everyone in the org will continue to have lgtm
on all repos. The idea is to fine-tune the list for some time
after this PR is merge, and then flip the switch.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>