Set default project permission to "none"
With this change, everyone in the Tekton org will stay part of the
org, but they won't be considered collaborators on a project unless
they're part of the project collaborator team.

The effectively restricts 'lgtm' to project collaborators only,
instead of being granted to everyone in the org for all repos like
it is today.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>