Remove Python 3.8 syntax
TIL about the := syntax which Python 3.8 introduced! The docs we have at
the moment say we require 3.6 (and I only have 3.7 installed personally
haha) so this commit removes the 3.8 syntax so that this script will
work with 3.6 (or at least 3.7 from my own attempts so far?).

I think in the long run we can simplify this by assuming folks will run
this via the image so they don't have to install it.