Add a TEP to start measuring Pipelines performance
This PR starts a TEP to begin to measure tekton pipelines performance
and address https://github.com/tektoncd/pipeline/issues/540

This first iteration just tries to describe the problem vs suggesting
the solution.

It DOES recommend measuring SLOs and SLIs as a goal, which is kind of
part of the solution, so if we think it's useful we could step back even
further, but I think this is a reasonable path forward, curious what
other folks think!