TEP 0030 Workspace Paths Technical Proposal
This commit adds a technical proposal to solve the problem statement laid out in
TEP 0030.

Workspace Paths allow a Task author to name specific important paths that they
require be provided on a Workspace or that they promise to produce onto a
Workspace. These paths are validated by Tekton to ensure that they exist before
or after Steps execute.