Spruce up user profiles 🌲
In the context of some discussions with @vdemeester, I've been working
on a doc and I wanted to be able to refer to something that
distinguishes between "runtime" and "authoring time". @jerop started
working on a doc for that and I thought it would be useful to be able to
refer to the 2 (potentially) different sets of users (users at authoring
time and users at runtime) - we already had this profiles doc and it
pretty much expressed that but I thought I'd try to make it a bit more
clear.

I also changed around the 'operator' bit - I'm not sure those personas
mentioned as being out of scope were actually fully out of scope since
we do need to support folks who are operating Tekton.

Added some links to this doc from other docs as well.