Authoring Time APIs vs Run Time APIs
We want to distinguish between authoring time APIs vs run time APIs to
facilitate discussions of whether a feature should be in Tasks/Pipelines
or TaskRuns/PipelineRuns -- this will also help with the reusability of
Tasks and Pipelines

At authoring time (i.e. when authoring Pipelines and Tasks), authors
should be able to include anything that is required for every execution
of the Task or Pipeline.

At run time (i.e. when invoking a Pipeline or Task via  PipelineRun or
TaskRun), users should be able to control execution as needed by their
context without having to modify Tasks and Pipelines.

Co-authored-by: Christie Wilson <christiewilson@google.com>