TEP-0045: WhenExpressions in Finally Tasks
Users can guard execution of `Tasks` using `WhenExpressions`,
but that is not supported in `Finally Tasks`.

This TEP describes the need for supporting `WhenExpressions` in
`Finally Tasks` not only to provide efficient guarded execution
but also to improve the reusability of `Tasks`.

Given we've recently added support for `Results` and `Status` in
`Finally Tasks`, this is an opportune time to enable `WhenExpressions`
in `Finally Tasks`.