Update Catalog Org TEP for Version Bump condition
This updates the TEP to in what case the version of the resource
should be updated in catalog. If there is a field addition with default
in existing resource and it is not going to change the behaviour of resource
and not gonna break existing pipelines then version bump is not required.

Signed-off-by: Shivam Mukhade <smukhade@redhat.com>