Add comparison with TEP-0046 🔎
In the most recent API working group we decided to keep this TEP and
[TEP-0046](https://github.com/tektoncd/community/pull/318) separate
because they are coming at a similar problem from different angles.

In the WG @jerop suggested that we update the TEPs with some info on
what the overlaps + differences are and that's what this TEP is adding!