Rename TEP to "Decouple Task Composition from Scheduling" ✍️
In today's API working group it was clear that I still haven't been able
to articulate what problem this TEP is trying to solve in a way that's
clear, so I'm trying this composition vs scheduling approach instead to
see if that is more clear. Also @skaegi pointed out that the overhead
required in running multiple pods is another consideration in addition
to the PVC overhead so I've tried to include this as well.

I'm going to follow up with a list of possible solutions so hopefully
that will help folks who are confused about what this is trying to
address.