Start brainstorming options 🧠
This commit adds possible solutions to the problem described in
TEP-0044, including references to solutions in other TEPS (46 & 54).
I was hoping to merge the problem statement before starting to talk
about solutions but it seems like the problem statement is too abstract
to get enough traction, and meanwhile folks have been opening more TEPs
with related proposals in the meantime, so hopefully starting to list
the options here will help us move the discussion forward.

I'm hoping we can merge the problem + possible options without needing
to decide on which one(s) we want to pursue.