Add proposal & alternatives for feature gates ⛩️
This commit adds a proposal for feature gates which has us using two
feature flags (alpha and beta) for gating accesss to fields as we add
them to beta and v1 CRDs, as well as using feature specific flags for
changes in behavior. It also tries to enumerate all the other ways we
could approach this in case something else stands out as more appealing.

I've tried to encorporate feedback from @vdemeester and @afrittoli
(via hackmd) as well but let me know if I missed any and/or ammend this
commit with more info.

This commit also transitions the proposal from proposed -> implementable
so that (once we agree and merge) we can stop blocking features in
Tekton Pipelines which are waiting on this.