TEP-0059: Skip Guarded Task Only
This TEP addresses skipping strategies to give users the flexibility to
skip a single guarded `Task` only and unblock execution of its dependent
`Tasks`.

We explore this separately from the [Conditions Beta TEP](https://github.com/tektoncd/community/blob/main/teps/0007-conditions-beta.md)
so that we can focus on this specific problem in further detail.

In this PR, we discuss the motivation, goals, non-goals and requirements
only so that we can align on the problem before exploring the many
alternatives later and choosing the way forward.