Update governing board and add instr for updates 🆕
@dibyom has joined the governing board after the most recent election,
@vdemeester and @bobcatfish have updated terms, and @ImJasonH was
removed (which was because he now works at RedHat which meant RedHat +
IBM had 3 seats which isn't allowed in our guidelines).

I've tried to add @dibyom to everything he needs to be added to and
wrote down how to do it - as well as removing @ImJasonH (which mostly
just removed his defunt google email XD)

Also added a former members section to give a shout out to folks who
used to be on the governing board!