Add alternate API WG 📅
We've added an additional API working group at a time that works for our
contributors in Australia; we've decided that since we have so much
content every week for the current API working group, instead of
alternating the times, we'll add an additional (but shorter) meeting
and see how it goes!

We're trying 6pm EST since this translated to 8am for the contributors
in Australia; any earlier is too early!

This commit also adds the folks as facilitators who indicated in the
survey that they were open to it. Thank you!!