TEP-0059: Skipping Strategies
In https://github.com/tektoncd/community/pull/388, we added the problem
statement for [TEP-0059: Skipping Strategies](https://github.com/tektoncd/community/blob/main/teps/0059-skip-guarded-task-only.md)
that addresses skipping strategies to give users the flexibility to skip
a single guarded `Task` only and unblock execution of its dependent
`Tasks`.

In this change, we add the proposal and discuss alternatives for solving
that problem.

Today, `WhenExpressions` are specified within `Tasks`, but they guard
the `Task` and its dependent `Tasks`.

To provide flexible skipping strategies, we propose changing the scope
of `WhenExpressions` from guarding a `Task` and its dependent `Tasks`
to guarding the `Task` only.

If a user wants to guard a `Task` and its dependent `Tasks`, they can:
1. cascade the `WhenExpressions` to the dependent `Tasks`
2. compose the `Task` and its dependent `Tasks` as a sub-`Pipeline`
that's guarded and executed together using `Pipelines` in `Pipelines`

```