TEP-0059: Skipping Strategies
In https://github.com/tektoncd/community/pull/388, we added the problem
statement for [TEP-0059: Skipping Strategies](https://github.com/tektoncd/community/blob/main/teps/0059-skip-guarded-task-only.md)
that addresses skipping strategies to give users the flexibility to skip
a single guarded `Task` only and unblock execution of its dependent
`Tasks`.

In this change, we add the proposal and discuss alternatives for solving
that problem.

Today, `WhenExpressions` are specified within `Tasks`, but they guard
the `Task` and its dependent `Tasks`.

To provide flexibility to either guard a `Task` only or guard a `Task`
and its dependent `Tasks`, we propose:
1. changing the scope of `WhenExpressions` to guard the `Task` only
2. cascading `WhenExpressions` to the dependent `Task` a user would
want to guard as well
3. using `Pipelines` in `Pipelines` to compose a set of `Tasks` that
should be guarded and executed together

```