Update TEP review process re `/approve` 🖃
[In our API working group on May 17](https://docs.google.com/document/d/17PodAxG8hV351fBhSu7Y_OIPhGTVgj6OJ2lPphYYRpU/edit#heading=h.ubgl0woia1ty)
we discussed how to improve our TEP review process such that it would
be possible to automate pinging folks when we are waiting on them to review
and generally how to make it more clear when a reviewer is expecting a
TEP author to make a change. We discussed the mechanics of approve vs
lgtm and decided that one improvement we could make would be for all
reviewers to use /approve - so there would be no need to add holds to
make sure a reviewer choosing to use /lgtm didnt abruptly merge the TEP.
This seems like a nice improvement, and it would be fairly easy to at
least create automation that is aware of who has and has not approved
the TEP among the reviewers (if we decide to go that far).