Remove myself as CLI maintainer 👻
I noticed I had been added as a reviewer to a CLI PR, and it was b/c I'm
in CLI maintainers. However I don't work on the CLI so I don't think I'm
in a good position to be reviewing PRs there, and so I'm removing myself
from the list. That leaves only @vdemeester as a CLI maintainer tho so I
can stay on the list if we wanted to make sure there was more than 1
person, just let me know.