TEP-0079: Tekton Catalog Support Tiers [Problem Statement]
This change adds the problem statement for Support Tiers in the Tekton
Catalog - we will explore the proposal in further changes.

The aim of this TEP is to establish support tiers for resources in the
Tekton Catalog to ensure users get high quality resources that they can
rely on while making it easy for contributors to submit resources to the
Tekton Catalog.

As discussed in [TEP-0003: Tekton Catalog Organization](https://github.com/tektoncd/community/blob/main/teps/0003-tekton-catalog-organization.md),
the support tiers are *commmunity*, *verified* and *official*, which
are differentiated as such:

|                        | Community |      Verified      |      Official      |
|:----------------------:|:---------:|:------------------:|:------------------:|
|   Automated Testing    |    :x:    | :heavy_check_mark: | :heavy_check_mark: |
| Images scanned for CVE |    :x:    |        :x:         | :heavy_check_mark: |
|  Maintained by Tekton  |    :x:    |        :x:         | :heavy_check_mark: |

Today, all the resources in the Tekton Catalog are in the *community*
support tier. To provide *verified* and *offical* support tiers, we
propose adding:

1. Automated testing of resources in the Tekton Catalog using Tekton
1. Regular scanning of used images for common vulnerabilities and exposures
1. Maintainance and ownership requirements for resources in the Tekton Catalog