TEP-0056: Pipelines in Pipelines - Add Use Case
In https://github.com/tektoncd/pipeline/issues/4067,
@mjgkaastrupandersen shares an additional use case for pipelines in
pipelines. In the issue, they describe how they need a gateway task or
grouping of pipelines to simplify specifying that a set of tasks all
need to wait for another set of tasks to complete execution. In this
change, we add that use case to [TEP-0056: Pipelines in Pipelines](https://github.com/tektoncd/community/blob/main/teps/0056-pipelines-in-pipelines.md).