TEP-0088: Tekton Results - Record Summaries
This TEP is a problem statement for introducing Record summaries to
Results in order to bubble up common/overall Result information without
needing to derive it from the underlying Records.

For now this TEP is only focusing on the problem statement to solicit
ideas. Detailed proposal to come in another PR.