TEP-0090: Matrix [Problem Statement]
This change adds the problem statement for Matrix. It scopes the
problem, describes the use cases, and identifies the goals, the
requirements for the solution, and related work in other continuous
delivery systems.

Today, users cannot supply varying `Parameters` to execute a
`PipelineTask`, that is, fan out a `PipelineTasks`.

To solve this problem, this TEP aims to enable executing the same
`PipelineTask` with different combinations of `Parameters` specified
in a `matrix`. `TaskRuns` or `Runs` will be created with variables
substituted with each combination of the `Parameters` in the `matrix`.

This `matrix` construct will enable users to specify concise but
powerful `Pipelines`. Moreover, it would improve the composability,
scalability, flexibility and reusability of *Tekton Pipelines*.

https://github.com/jerop/community/blob/looping/teps/0090-matrix.md

References:
- [Task Loops Experimental Project][task-loops]
- Issues:
  - https://github.com/tektoncd/pipeline/issues/2050
  - https://github.com/tektoncd/pipeline/issues/4097

[task-loops]: https://github.com/tektoncd/experimental/tree/main/task-loops