Add code of conduct to lists gov board members should join 📝
I just recently realized that only @vdemeester and I were on the code of
conduct mailing list - even though being responsible for this is listed
as a duty for the entire governing board. After checking with the board
I've added everyone to the list and I'm recording it here so we don't
forget in future as we make adjustments to the board.

(Also fixed typo in tekton-users link.)