Add contributor ladder.
This is largely based off of
https://github.com/cncf/project-template/blob/main/CONTRIBUTOR_LADDER.md.

This helps define more lower-stakes roles for getting involved with Tekton, as
well as updates existing roles to provide expectations of the roles.
Additionally, this adds a formal process to revoke permissions in
instances of inactivity or abuse.

I tried my best to keep the requirements the same as before, so this
shouldn't be introducing any new requirements for the
collaborator/maintainer roles.