Require TEP-0044 before removing PipelineResources 🙋‍♀️
The Google Tekton team was discussing plans for 2022 and we were trying
to figure out how important pipeline in a pod was - based on the initial
requirements here for PipelineResource deprecation, it felt like we'd be
able to go ahead (specifically with Pipelines v1!) without pipeline in a
pod. The existing requirements said we'd need experimental support to
prove it was possible - however when thinking through what that would
mean for v1, i.e.: pipelineresources are now gone, and you need to use
this experimental custom task if you want the same "in a pod"
functionality - that didn't seem fair to users.

So this change would update our requirements so that we'd have at least
alpha support for TEP-0044 (task and pipeline decoupled, possible as a
pipeline in a pod)  functionality before we'd mark
this TEP as implementable, i.e. before actually removing any
PipelineResource functionality from our API surface.