TEP-0095 Add a problem statement for common repository configuration
We have a number of projects in-flight that refer to or directly use
repository information. It might be a good idea to try and introduce
a single place for configuration of this kind of info to avoid a
situation where users / operators have to individually configure
multiple different components with the same repo information.

This commit adds a problem statement and use-cases for adding a
repository object to Tekton's data model to try and address this
potential issue.