[TEP-0096]: Pipelines V1 Scope
This TEP adds use cases for Tekton and criteria for scoping
what work should and should not be prioritized for inclusion in a V1 API.
It also adds a stability policy for metrics.

Specific features, and whether they will be in scope for V1, will
be proposed in a subsequent commit.

Co-authored-by: Jerop Kipruto jerop@google.com
Co-authored-by: Priti Desai pdesai@us.ibm.com