[TEP-0012]: Mark as implemented
Tekton now has an API spec and conformance tests for TaskRun.
We may want to add conformance docs and tests for other parts of the Tekton API,
but the foundation has been laid for adding these later,
and this work can be tracked through Github issues instead of leaving this TEP open.