Add governing board / community meeting 👩‍⚖️
The governing board sync meeting has so far been not publicly advertised
(although the notes are available to the community via tekton-dev and
tekton-users). Recently we discussed potentially creating a "community
working group", however it felt like that most closely resembles what we
discuss in governing board meetings so instead of adding a new meeting,
with thought it might make sense to make this existing governing board
meeting public and see if other ppl are interested in joining and if
this can meet the need.

We'd also like to encourage more people to run for the governing board
(esp with the upcoming election in feb!) and joining this meeting could
give ppl an idea of what's involved.