TEP-0079: Tekton Catalog Support Tiers [Problem Statement]
This change adds the problem statement for Support Tiers in the Tekton
Catalog - we will explore the proposal in further changes.

The aim of this TEP is to establish support tiers for resources in the
Tekton Catalog to ensure users get high quality resources that they can
rely on while making it easy for contributors to submit resources to the
Tekton Catalog.

The goals of the TEP are to address the following aspects of the Tekton
Catalog:
- Ownership and Maintenance
- Automated Testing and Dogfooding
- Image Scanning for Common Vulnerabilities and Exposures (CVEs)
- Verified Remote Resources

Related PRs:
- https://github.com/tektoncd/community/pull/494
- https://github.com/tektoncd/community/pull/170
- https://github.com/tektoncd/community/pull/495