TEP-0079: Tekton Catalog Support Tiers [Proposal - Ownership and Maintenance]
In [TEP-0079: Tekton Catalog Support Tiers][tep-0079], we have defined
the problem statement for establishing support tiers for resources in
the Tekton Catalog.

In this change, we address the first goal of [Ownership and Maintenance][goal].

As previously discussed in [TEP-0003][tep-0003-upstream], we propose creating
two support tiers, `Community` and `Official`, through separate Tekton Catalogs.

The Community Catalog would make it easy for contributors to share resources,
while the Official Catalog will be a corpus of high quality resources that
users can rely on. This distribution of ownership would ensure the workload
is sustainable.

The other goals and details of testing, scanning and verification will be
explored in the subsequent changes.

[tep-0079]: https://github.com/tektoncd/community/blob/main/teps/0079-tekton-catalog-support-tiers.md
[goal]: https://github.com/tektoncd/community/blob/main/teps/0079-tekton-catalog-support-tiers.md#ownership-and-maintenance
[tep-0003-upstream]: https://github.com/tektoncd/community/blob/main/teps/0003-tekton-catalog-organization.md#upstream-catalogs