TEP-0090: Matrix [Proposal]
In [TEP-0090: Matrix][tep-0090] we described the problem
statement for fanning out `PipelineTasks`. That is,
enabling execution of a `PipelineTask` with different
permutations of `Parameters` specified in a `Matrix`.

In this change, we add the proposal to solve the above problem.
It includes the design details, implementation plan, promotion
plan, design evalutation, future work, alternatives, and others.

In summary, we propose adding a `matrix` field to the `PipelineTask`
specification that will be used to declare `Parameters` of type `Array`.
The `PipelineTask` will be executed in parallel `TaskRuns` or `Runs`
with its `Parameters` substituted with the permutations of
`Parameters` in the `Matrix`.

[tep-0090]: https://github.com/tektoncd/community/blob/main/teps/0090-matrix.md